be silly use unsigned datatypes :3

+++ ClientPhotoresistor/ClientPhotoresistor.ino
@@ -14,7 +14,7 @@
 int RIGHT_CUTOFF;
 int LEFT_CUTOFF;
 
-int nextUpdate = UPDATE_FREQUENCY;
+unsigned long nextUpdate = UPDATE_FREQUENCY;
 int currentLives = 0;
 SegmentDisplay segmentDisplay(4, 5, 10, 11, 9, 8, 7, 6);
